Reorganize function code
